Better organize files and update requires
