+++ lib/active_replicas/proxying_connection_pool.rb
@@ -1,5 +1,6 @@
-require 'monitor'
+require 'active_support/core_ext/module/delegation'
 require 'active_support/hash_with_indifferent_access'
+require 'monitor'
 
 module ActiveReplicas
   # Manages connection pools to the primary and replica databases. Returns
